Split GenericContainer tests to run faster
